# backend/muni-valle-sol-bff/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Municipalidad Valle del Sol - BFF API
  description: |
    Backend For Frontend (BFF) para la plataforma de gestión de incendios
    de la Municipalidad Valle del Sol.
    
    <p>Características principales:</p>
    <ul>
      <li>Autenticación JWT mediante cookies HttpOnly</li>
      <li>Gestión de reportes de incendios en tiempo real</li>
      <li>Sistema de alertas automáticas según severidad</li>
      <li>Integración con microservicios especializados</li>
      <li>Resilencia mediante Circuit Breaker (Resilience4j)</li>
    </ul>
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: desarrollo@munivallesol.cl
    url: https://munivallesol.cl
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: https://munivallesol.cl/logo.png
    altText: Municipalidad Valle del Sol
servers:
  - url: http://localhost:8080
    description: Servidor Local
    variables:
      basePath:
        default: /api
  - url: https://bff-dev.munivallesol.cl
    description: Ambiente Desarrollo
    variables:
      basePath:
        default: /api
  - url: https://bff-prod.munivallesol.cl
    description: Ambiente Producción
    variables:
      basePath:
        default: /api
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        Token JWT almacenado en cookie HttpOnly.
        Se establece automáticamente tras login.
  schemas:
...
    Location:
      type: object
      properties:
        lat:
          type: number
          format: double
          example: -33.8688
          description: Latitud geográfica
        lng:
          type: number
          format: double
          example: -71.5203
          description: Longitud geográfica
      description: Coordenadas geográficas del incidente
    ReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
          description: Identificador único del reporte
        titulo:
          type: string
          example: Incendio en Cerro Norte
          description: Título descriptivo del reporte
        descripcion:
          type: string
          example: Fuego activo en ladera norte, cerca de viviendas
          description: Descripción detallada del incidente
        tipo:
          type: string
          enum: [INCENDIO, HUMO, SOSPECHOSO]
          example: INCENDIO
          description: Clasificación del incidente
        estado:
          type: string
          enum: [ACTIVO, RESUELTO, CERRADO]
          example: ACTIVO
          description: Estado actual del reporte
        emailUsuario:
          type: string
          format: email
          example: juan@example.com
          description: Email del usuario que reportó
        ubicacion:
          $ref: '#/components/schemas/Location'
        fechaCreacion:
          type: string
          format: date-time
          example: 2026-06-13T14:30:00
          description: Fecha y hora de creación del reporte
      description: Reporte de incendio con ubicación anidada
    CreateReportRequest:
      type: object
      required:
        - titulo
        - descripcion
        - tipo
        - emailUsuario
        - latitud
        - longitud
      properties:
        titulo:
          type: string
          example: Incendio en Cerro Norte
          description: Título descriptivo
        descripcion:
          type: string
          example: Fuego activo en ladera norte
          description: Descripción del incidente
        tipo:
          type: string
          enum: [INCENDIO, HUMO, SOSPECHOSO]
          example: INCENDIO
          description: Tipo de incidente
        emailUsuario:
          type: string
          format: email
          example: juan@example.com
          description: Email del usuario que reporta
        latitud:
          type: number
          format: double
          example: -33.8688
          description: Coordenada de latitud
        longitud:
          type: number
          format: double
          example: -71.5203
          description: Coordenada de longitud
      description: Solicitud para crear un nuevo reporte
    UpdateStatusRequest:
      type: object
      required:
        - estado
      properties:
        estado:
          type: string
          enum: [ACTIVO, RESUELTO, CERRADO]
          example: RESUELTO
          description: Nuevo estado del reporte
      description: Solicitud para actualizar estado de reporte
    UpdateTitleRequest:
      type: object
      required:
        - titulo
      properties:
        titulo:
          type: string
          example: Incendio Zona Crítica - ACTUALIZADO
          description: Nuevo título del reporte
      description: Solicitud para actualizar título de reporte

    # ──── DTOs de Alertas ────
...
paths:
...
  /reportes:
    get:
      tags:
        - Reportes
      summary: Lista todos los reportes
      description: |
        Obtiene la lista completa de reportes sin filtros.
        Incluye reportes en todos los estados (ACTIVO, RESUELTO, CERRADO).
        
        **Circuit Breaker:** Si MS-Reportes no responde, retorna lista vacía.
      operationId: listAllReports
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Lista de reportes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportDTO'
        '401':
          description: Usuario no autenticado

    post:
      tags:
        - Reportes
      summary: Crea un nuevo reporte de incendio
      description: |
        Registra un nuevo incendio en el sistema.
        
        **Flujo:**
        1. Se validan todos los campos obligatorios
        2. Se transforma el record a Map en ReportRequestHandler
        3. Se envía a MS-Reportes para persistencia
        4. Se mapea la respuesta a ReportDTO
        
        **Tipos de incidente:**
        - INCENDIO: Fuego activo (severidad → ALTA)
        - HUMO: Presencia de humo (severidad → MEDIA)
        - SOSPECHOSO: Actividad sospechosa (severidad → BAJA)
      operationId: createReport
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportRequest'
            example:
              titulo: Incendio en Cerro Norte
              descripcion: Fuego activo en ladera norte, cerca de viviendas
              tipo: INCENDIO
              emailUsuario: juan@example.com
              latitud: -33.8688
              longitud: -71.5203
      responses:
        '201':
          description: Reporte creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDTO'
        '400':
          description: Datos inválidos
        '401':
          description: Usuario no autenticado
        '503':
          description: MS-Reportes no disponible
  /reportes/activos:
    get:
      tags:
        - Reportes
      summary: Lista reportes activos
      description: |
        Obtiene solo los reportes con estado ACTIVO.
        Útil para dashboard en tiempo real y alertas.
      operationId: listActiveReports
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Lista de reportes activos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportDTO'
        '401':
          description: Usuario no autenticado
  /reportes/{id}:
    get:
      tags:
        - Reportes
      summary: Obtiene un reporte por ID
      description: Recupera los detalles completos de un reporte específico.
      operationId: getReportById
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 101
          description: Identificador del reporte
      responses:
        '200':
          description: Reporte encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDTO'
        '401':
          description: Usuario no autenticado
        '404':
          description: Reporte no encontrado
        '503':
          description: MS-Reportes no disponible

    put:
      tags:
        - Reportes
      summary: Actualiza el título de un reporte
      description: Modifica el título descriptivo de un reporte existente.
      operationId: updateReportTitle
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTitleRequest'
            example:
              titulo: Incendio Zona Crítica - ACTUALIZADO
      responses:
        '200':
          description: Reporte actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDTO'
        '401':
          description: Usuario no autenticado
        '404':
          description: Reporte no encontrado
        '503':
          description: MS-Reportes no disponible

    delete:
      tags:
        - Reportes
      summary: Elimina un reporte
      description: Borra permanentemente un reporte del sistema.
      operationId: deleteReport
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 101
      responses:
        '204':
          description: Reporte eliminado exitosamente
        '401':
          description: Usuario no autenticado
        '404':
          description: Reporte no encontrado
        '503':
          description: MS-Reportes no disponible
  /reportes/{id}/estado:
    put:
      tags:
        - Reportes
      summary: Actualiza el estado de un reporte
      description: |
        Cambia el estado de un reporte.
        
        **Estados disponibles:**
        - ACTIVO: Incidente en curso
        - RESUELTO: Incidente controlado/extinguido
        - CERRADO: Incidente completamente resuelto
      operationId: updateReportStatus
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 101
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
            examples:
              resolved:
                summary: Marcar como resuelto
                value:
                  estado: RESUELTO
              closed:
                summary: Cerrar completamente
                value:
                  estado: CERRADO
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDTO'
        '401':
          description: Usuario no autenticado
        '404':
          description: Reporte no encontrado
        '503':
          description: MS-Reportes no disponible

  # ══════════════════════════════════════════════════════════════════════════════
  # ALERTAS
  # ══════════════════════════════════════════════════════════════════════════════
...
security:
  - cookieAuth: []